Move DataSource dependency detection into dedicated package
Closes gh-25757
