+++ spring-boot-project/spring-boot-autoconfigure/src/main/java/org/springframework/boot/autoconfigure/jdbc/DataSourceInitializationDataSourceInitializerDetector.java
@@ -19,8 +19,8 @@ package org.springframework.boot.autoconfigure.jdbc;
 import java.util.Collections;
 import java.util.Set;
 
-import org.springframework.boot.jdbc.init.AbstractBeansOfTypeDataSourceInitializerDetector;
-import org.springframework.boot.jdbc.init.DataSourceInitializerDetector;
+import org.springframework.boot.jdbc.init.dependency.AbstractBeansOfTypeDataSourceInitializerDetector;
+import org.springframework.boot.jdbc.init.dependency.DataSourceInitializerDetector;
 
 /**
  * A {@link DataSourceInitializerDetector} for {@link DataSourceInitialization}.
