+++ spring-boot-project/spring-boot-autoconfigure/src/main/resources/META-INF/spring.factories
@@ -168,7 +168,6 @@ org.springframework.boot.autoconfigure.thymeleaf.ThymeleafTemplateAvailabilityPr
 org.springframework.boot.autoconfigure.web.servlet.JspTemplateAvailabilityProvider
 
 # DataSource initializer detectors
-org.springframework.boot.jdbc.init.DataSourceInitializerDetector=\
+org.springframework.boot.jdbc.init.dependency.DataSourceInitializerDetector=\
 org.springframework.boot.autoconfigure.flyway.FlywayMigrationInitializerDataSourceInitializerDetector,\
 org.springframework.boot.autoconfigure.jdbc.DataSourceInitializationDataSourceInitializerDetector
-
