+++ spring-boot-project/spring-boot/src/main/java/org/springframework/boot/jdbc/init/dependency/AnnotationDependsOnDataSourceInitializationDetector.java
@@ -14,39 +14,30 @@
  * limitations under the License.
  */
 
-package org.springframework.boot.jdbc.init;
+package org.springframework.boot.jdbc.init.dependency;
 
-import java.util.Collections;
+import java.util.HashSet;
 import java.util.Set;
-
-import javax.sql.DataSource;
 
 import org.springframework.beans.factory.config.ConfigurableListableBeanFactory;
 
 /**
- * Base class for {@link DataSourceInitializerDetector DataSourceInitializerDetectors}
- * that detect {@link DataSource} beans by type.
+ * {@link DependsOnDataSourceInitializationDetector} that detects beans annotated with
+ * {@link DependsOnDataSourceInitialization}.
  *
  * @author Andy Wilkinson
- * @since 2.5.0
  */
-public abstract class AbstractBeansOfTypeDataSourceInitializerDetector implements DataSourceInitializerDetector {
-
-	/**
-	 * Returns the bean types that should be detected as being data source initializers.
-	 * @return the data source initializer bean types
-	 */
-	protected abstract Set<Class<?>> getDataSourceInitializerBeanTypes();
+class AnnotationDependsOnDataSourceInitializationDetector implements DependsOnDataSourceInitializationDetector {
 
 	@Override
 	public Set<String> detect(ConfigurableListableBeanFactory beanFactory) {
-		try {
-			Set<Class<?>> types = getDataSourceInitializerBeanTypes();
-			return new BeansOfTypeDetector(types).detect(beanFactory);
+		Set<String> dependentBeans = new HashSet<>();
+		for (String beanName : beanFactory.getBeanDefinitionNames()) {
+			if (beanFactory.findAnnotationOnBean(beanName, DependsOnDataSourceInitialization.class) != null) {
+				dependentBeans.add(beanName);
+			}
 		}
-		catch (Throwable ex) {
-			return Collections.emptySet();
-		}
+		return dependentBeans;
 	}
 
 }
