+++ spring-boot-project/spring-boot/src/main/java/org/springframework/boot/jdbc/init/dependency/package-info.java
@@ -15,6 +15,6 @@
  */
 
 /**
- * General infrastructure for {@code DataSource} initialization.
+ * Infrastructure for establishing {@code DataSource} initialization bean dependencies.
  */
-package org.springframework.boot.jdbc.init;
+package org.springframework.boot.jdbc.init.dependency;
